Splits CLI commands into individual packages
